New contract resolver functionality
Added functionality to create custom contract resolvers.
