+++ src/WebApiContrib.Formatting.Xlsx/XlsxMediaTypeFormatter.cs
@@ -1,14 +1,10 @@
-﻿using OfficeOpenXml;
-using OfficeOpenXml.Style;
+﻿using OfficeOpenXml.Style;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Security.Permissions;
 using System.Threading.Tasks;
-using System.Web.ModelBinding;
 using util = WebApiContrib.Formatting.Xlsx.FormatterUtils;
 using WebApiContrib.Formatting.Xlsx.Attributes;
 using WebApiContrib.Formatting.Xlsx.Serialisation;
